move the “GameConsts” file to the “lib” folder at the root of the project
