+++ app/services/createRoom.js
@@ -4,7 +4,7 @@ const moment = require('moment')
 const _ = require('lodash')
 const hri = require('human-readable-ids').hri
 
-const GameConsts = require('../../resources/assets/js/lib/GameConsts')
+const GameConsts = require('../../lib/GameConsts')
 const getTeam = require('./getTeam')
 
 module.exports = function(data) {
