+++ nodemon.json
@@ -14,7 +14,7 @@
     "bin/",
     "app/",
     "app.js",
-    "resources/assets/js/lib/GameConsts.js"
+    "lib/GameConsts.js"
   ],
   "env": {
     "NODE_ENV": "development"
