+++ resources/assets/js/lib/RemainingFuelPercent.js
@@ -1,4 +1,4 @@
-import GameConsts from './GameConsts'
+import GameConsts from 'lib/GameConsts'
 
 export default function RemainingFuelPercent(counter) {
     const maxFuel = GameConsts.JUMP_JET_STARTING_FUEL * -1
