+++ resources/assets/js/lib/SocketEvents/onPlayerRespawn.js
@@ -4,7 +4,7 @@ import includes from 'lodash/includes'
 import Maps from '../Maps'
 import PlayerById from '../PlayerById'
 import actions from '../../actions'
-import GameConsts from '../GameConsts.js'
+import GameConsts from 'lib/GameConsts'
 import CreateKeyboardBindings from '../CreateHandler/CreateKeyboardBindings'
 import GetSpawnPoint from '../GetSpawnPoint'
