+++ resources/assets/js/reducers/game.js
@@ -1,5 +1,5 @@
 import storage from 'store'
-import GameConsts from '../lib/GameConsts'
+import GameConsts from 'lib/GameConsts'
 
 const initialState = {
     chatModalIsOpen: false,
