+++ resources/assets/js/states/TeamDeathmatch.js
@@ -7,7 +7,7 @@ import FireRocket from '../lib/FireRocket'
 import RotateBulletsToTrajectory from '../lib/RotateBulletsToTrajectory'
 import Maps from '../lib/Maps'
 import actions from '../actions'
-import GameConsts from '../lib/GameConsts'
+import GameConsts from 'lib/GameConsts'
 import UpdateHudPositions from '../lib/UpdateHudPositions'
 import UpdateHurtBorder from '../lib/UpdateHurtBorder'
 import UpdatePlayerPosition from '../lib/UpdatePlayerPosition'
