+++ resources/assets/js/ui/components/SettingsModal/SettingsView.js
@@ -2,7 +2,7 @@ import React, { PropTypes } from 'react'
 import autobind from 'react-autobind'
 
 import NameGenerator from '../../../lib/NameGenerator'
-import GameConsts from '../../../lib/GameConsts'
+import GameConsts from 'lib/GameConsts'
 
 export default class SettingsView extends React.Component {
     constructor(props) {
